Fixed minor compilation warning on casting double to unsigned integer by adding explicit cast. Not reported on bug tracker.

+++ LatBo/GridObj_ops_immersedboundary.cpp
@@ -90,7 +90,7 @@ double GridObj::ibm_deltakernel(double radius, double dilation) {
 void GridObj::ibm_findsupport(unsigned int ib, unsigned int m) {
 
 	// Declarations
-	size_t inear, jnear;						// Nearest node indices
+	unsigned int inear, jnear;					// Nearest node indices
 	double dist_x, dist_y, delta_x, delta_y;	// Distances and deltas
 #if (dims == 3)
 	// Extras for 3D
@@ -100,8 +100,8 @@ void GridObj::ibm_findsupport(unsigned int ib, unsigned int m) {
 
 
 #ifdef CHEAP_NEAREST_NODE_DETECTION
-	inear = floor( iBody[ib].markers[m].position[0]/dx );
-	jnear = floor( iBody[ib].markers[m].position[1]/dy );
+	inear = (unsigned int)floor( iBody[ib].markers[m].position[0]/dx );
+	jnear = (unsigned int)floor( iBody[ib].markers[m].position[1]/dy );
 
 #if (dims == 3)
 	knear = floor( iBody[ib].markers[m].position[2]/dz );
